fixed the insight code
Now it can be used seamlessly with any screen resolutions given the calibration spec.

Fixed the problem of multiple datapoint assosiations by incrementing the entry of sliding window from the point of hit occurence.

The distance threshold for hit classification is now based on foveal_diameter. Threshold is calculated using the avg values of human eye_ball radius, distance to the screen and foveal diamter